modified:   rsc/scripts/main.sh

+++ rsc/scripts/main.sh
@@ -10,6 +10,7 @@ echo "---------------------------"
 
 # start in background mode ?
 if [ x"$1" = "xbg" ] ; then
+  if [ -d $(dirname $0)/.lockdir121978 ] ; then echo "$0 : --> another instance is already running - exiting." ; exit 1 ; fi
   trap 'echo -e "\n********************\nTo cancel background job type \n kill $(cat ./log | grep Job | cut -d : -f 2)\n\nTo view logfile type \n tail -f log\n"********************' EXIT
   cd $(dirname $0)
   nohup ./run_script.sh &> log &
